Delete RCs in the integration tests by label, not name
The name can change (e.g. in Kubernetes 1.0.1 it's "monitoring-heapster-v5"
while it used to be "monitoring-heapster") so it is simply not good enough to
find the correct RCs.

Defaulting to "heapster", we take the value of either "k8s-app" or "name" from
the given RC config labels. Then we find exactly one RC that matches either
k8s-app=heapster or name=heapster and we proceed to delete it.