Don't gitignore generated files
`go generate` should be run by the developers and the files commited to the
repository. The user shouldn't need to run go generate before building.

From the `go generate` design doc:

First, go generate is intended to be run by the author of a package, not the
client of it. The author of the package generates the required Go files and
includes them in the package; the client does a regular go get or go build.
Generation through go generate is not part of the build, just a tool for
package authors. This avoids complicating the dependency analysis done by Go
build.