Oldtimer: Introduce Historical API
Oldtimer is the Heapster historical metrics access mechanism.  This
commit introduces the the API for Oldtimer, which mirrors the Heapster
model API, with additional paths for retrieving metrics aggregated over
time.

In order for the historical API to function, one specified sink must
support the historical access interface defined in this commit.