Push Metrics: allow limit for number of metrics
This commit introduces a flag (`--max_push_metrics`) which limits
the number of allowed unique metric names per push source.  By
default, the flag is set to zero, which means an unlimitted number
are allowed.  The set of metrics pushed by each source is remembered
for the lifetime of the heapster process.