Push Metrics: Properly merge labeled metrics
Previously, labeled metrics were simply appended without regards for
label set.  This commit changes it so that two labeled metrics with
idential names and label sets are treated the same as two metrics with
identical names, when inside the push handler.

It also refactors the push handler types in order to properly separately
track the push count for averaging gauge metrics (instead of using the
unused field in the normal MetricValue).