Add additional logging around points of failure
This commit adds additional logging in the rate calculator and summary
source around potential points of failure.  Namely, additional logging
was introduced when the rate calculator skips calculation, and the
summary source was enhanced to log additional information about fetching
and processing data from the Kubelet at levels 9 and 10 (the latter of
which simply logs the entire response from the Kubelet).